requirements.txt
torch>=1.9.0
torchvision>=0.10.0
diffusers>=0.10.2
numpy>=1.21.0
Pillow>=8.3.0
tqdm>=4.62.0
matplotlib>=3.4.0
